ASoC: samsung: odroid: Fix EPLL frequency values
To prevent incorrect setting of the EPLL the clock frequency
values are changed to exact values as possible to obtain on
the EPLL output with given PLL coefficients.
This patch is required after recent change of the EPLL rate
table by patch
"clk: samsung: exynos5420: The EPLL rate table corrections".

Signed-off-by: Sylwester Nawrocki <s.nawrocki@samsung.com>
---
It would be good to apply this patch for 4.13-rc, I plan to
get applied corresponding clk patch also to v4.13-rc.